+++ src/backend/defs/go-objects/nats.go
@@ -19,16 +19,16 @@ type IMAPorder struct {
 }
 
 type BrokerOrder struct {
-	MessageId string `json:"message_id"`
-	Order     string `json:"order"`
-	RemoteId  string `json:"remote_id"`
-	UserId    string `json:"user_id"`
+	MessageId  string `json:"message_id"`
+	Order      string `json:"order"`
+	IdentityId string `json:"identity_id"`
+	UserId     string `json:"user_id"`
 }
 
 // DeliveryAck holds reply from nats when using request/reply system for messages
 type DeliveryAck struct {
-	Err          bool          `json:"error"`
-	Response     string        `json:"message,omitempty"`
+	Err      bool   `json:"error"`
+	Response string `json:"message,omitempty"`
 }
 
 // a lighter struct to reply to a request